feat: explore dialog watcher
