+++ gui/src/util/localStorage.ts
@@ -3,6 +3,7 @@ import { OnboardingStatus } from "../components/OnboardingCard";
 
 type LocalStorageTypes = {
   isExploreDialogOpen: boolean;
+  hasDismissedExploreDialog: boolean;
   onboardingStatus?: OnboardingStatus;
   hasDismissedOnboardingCard: boolean;
   mainTextEntryCounter: number;
@@ -17,6 +18,11 @@ type LocalStorageTypes = {
   /** Indicates whether the user has seen any kind of onboarding card since hub.continue.dev */
   seenHubIntro: boolean;
 };
+
+export enum LocalStorageKey {
+  IsExploreDialogOpen = "isExploreDialogOpen",
+  HasDismissedExploreDialog = "hasDismissedExploreDialog",
+}
 
 export function getLocalStorage<T extends keyof LocalStorageTypes>(
   key: T,
